Large update
Sorry for the big commit without much nuance. I had forgotten to make earlier commits,
and a lot of changes were rolled in here. In general, this continues to make minor changes to improve
the provisioning workflow, the API, and also the web front-end updates were taking place
during this commits, so updates to make the two compatible again are included as well.

Signed-off-by: Matt Oswalt <matt@oswalt.dev>