Initial work on OpenTracing instrumentation
This is a good early first effort, but I'm learning more as I go, so I'm
not finished yet. The way I'm starting spans could be simplified, and I haven't
yet converted logging over to span event logs. But I wanted to make this commit
before I go further since it works, just not in the end-state

Signed-off-by: Matt Oswalt <matt@oswalt.dev>