Fixing the remainder of the txn bugs I could find
These are in yet-unused functions, but still good to fix. I had another PR open
to fix these, but accidentally branched off of this to move to go modules,
and I'm too lazy to go back and fix the history as well as move directories back to
GOPATH to make this fix, so here we are.

Signed-off-by: Matt Oswalt <matt@oswalt.dev>